Move subcommand tests to their own file and remove duplicate tests
